+++ tests/QRegexSearchTests.cpp
@@ -26,6 +26,10 @@ private slots:
     void dotMatchesNewlineFlagIsHonored();
     void posixCharacterClassesRemainSupported();
     void utf8PositionsRemainScintillaByteOffsets();
+    void patternLengthIsExplicit();
+    void replacementSupportsBoostStyleReferences();
+    void replacementSupportsEscapedCharacters();
+    void invalidPatternsClearMatchLength();
     void replacementExpandsCaptureGroups();
 };
 
@@ -89,6 +93,7 @@ void QRegexSearchTests::dotMatchesNewlineFlagIsHonored()
     QCOMPARE(search.FindText(&document, 0, document.Length(), "a.b", true, false, false,
                              FindOption::RegExp, &length), Sci::Position(-1));
 
+    length = 3;
     const Sci::Position position = search.FindText(&document, 0, document.Length(), "a.b",
                                                     true, false, false,
                                                     FindOption::RegExp | FindOption::Cxx11RegEx,
@@ -102,10 +107,11 @@ void QRegexSearchTests::posixCharacterClassesRemainSupported()
     Document document(DocumentOption::Default);
     setDocumentText(document, "id=42");
     QRegexSearch search;
-    Sci::Position length = 1;
+    const QByteArray pattern = QByteArrayLiteral("[[:digit:]]+");
+    Sci::Position length = pattern.size();
 
-    const Sci::Position position = search.FindText(&document, 0, document.Length(),
-                                                    "[[:digit:]]+", true, false, false,
+    const Sci::Position position = search.FindText(&document, 0, document.Length(), pattern.constData(),
+                                                    true, false, false,
                                                     FindOption::RegExp | FindOption::Posix,
                                                     &length);
 
@@ -128,14 +134,78 @@ void QRegexSearchTests::utf8PositionsRemainScintillaByteOffsets()
     QCOMPARE(length, Sci::Position(3));
 }
 
+void QRegexSearchTests::patternLengthIsExplicit()
+{
+    Document document(DocumentOption::Default);
+    setDocumentText(document, "prefix foo suffix");
+    QRegexSearch search;
+    Sci::Position length = 3;
+    const QByteArray pattern("foo\0ignored", 11);
+
+    const Sci::Position position = search.FindText(&document, 0, document.Length(), pattern.constData(),
+                                                    true, false, false, FindOption::RegExp, &length);
+
+    QCOMPARE(position, Sci::Position(7));
+    QCOMPARE(length, Sci::Position(3));
+}
+
+void QRegexSearchTests::replacementSupportsBoostStyleReferences()
+{
+    Document document(DocumentOption::Default);
+    setDocumentText(document, "foo-bar");
+    QRegexSearch search;
+    const QByteArray pattern = QByteArrayLiteral("(?<left>foo)-(bar)");
+    Sci::Position length = pattern.size();
+
+    QCOMPARE(search.FindText(&document, 0, document.Length(), pattern.constData(), true, false, false,
+                             FindOption::RegExp, &length), Sci::Position(0));
+
+    const QByteArray replacement = QByteArrayLiteral("${left}/$2/$0/$MATCH/\\1");
+    length = replacement.size();
+    const char *result = search.SubstituteByPosition(&document, replacement.constData(), &length);
+
+    QCOMPARE(QByteArray(result, length), QByteArray("foo/bar/foo-bar/foo-bar/foo"));
+}
+
+void QRegexSearchTests::replacementSupportsEscapedCharacters()
+{
+    Document document(DocumentOption::Default);
+    setDocumentText(document, "foo");
+    QRegexSearch search;
+    const QByteArray pattern = QByteArrayLiteral("(foo)");
+    Sci::Position length = pattern.size();
+
+    QCOMPARE(search.FindText(&document, 0, document.Length(), pattern.constData(), true, false, false,
+                             FindOption::RegExp, &length), Sci::Position(0));
+
+    const QByteArray replacement = QByteArrayLiteral("\\1\\n\\t\\\\");
+    length = replacement.size();
+    const char *result = search.SubstituteByPosition(&document, replacement.constData(), &length);
+
+    QCOMPARE(QByteArray(result, length), QByteArray("foo\n\t\\"));
+}
+
+void QRegexSearchTests::invalidPatternsClearMatchLength()
+{
+    Document document(DocumentOption::Default);
+    setDocumentText(document, "foo");
+    QRegexSearch search;
+    Sci::Position length = 4;
+
+    QCOMPARE(search.FindText(&document, 0, document.Length(), "(", true, false, false,
+                             FindOption::RegExp, &length), Sci::Position(-1));
+    QCOMPARE(length, Sci::Position(0));
+}
+
 void QRegexSearchTests::replacementExpandsCaptureGroups()
 {
     Document document(DocumentOption::Default);
     setDocumentText(document, "foo");
     QRegexSearch search;
-    Sci::Position length = 3;
+    const QByteArray pattern = QByteArrayLiteral("(foo)");
+    Sci::Position length = pattern.size();
 
-    QCOMPARE(search.FindText(&document, 0, document.Length(), "(foo)", true, false, false,
+    QCOMPARE(search.FindText(&document, 0, document.Length(), pattern.constData(), true, false, false,
                              FindOption::RegExp, &length), Sci::Position(0));
 
     const QByteArray replacement = "\\1!";